\name{annotate.data}
\alias{annotate.data}
\title{Annotate Differentially expressed gene table}
\usage{
  annotate.data(datTable, anot.file, source.format = "gff",
    SelFeature = "gene",
    fields = c("ID", "Name", "description"),
    cleandescription)
}
\arguments{
  \item{datTable}{DE results table}

  \item{anot.file}{}

  \item{source.format}{}

  \item{SelFeature}{}

  \item{fields}{}

  \item{cleandescription}{}
}
\value{
  annotated DE genes results table
}
\description{
  Annotate Differentially expressed gene table
}

