\name{ascii2char}
\alias{ascii2char}
\title{clean up description by converting ascii code to char in the annotation}
\usage{
  ascii2char(myanot, field)
}
\arguments{
  \item{myanot}{a dataframe containing results}

  \item{field}{name of the field that require conversion}
}
\value{
  a dataframe with cleaned field
}
\description{
  clean up description by converting ascii code to char in
  the annotation
}

