\name{plot.cluster2}
\alias{plot.cluster2}
\title{Plot heatmap and 2D scatter plot of principal components}
\usage{
  plot.cluster2(CountsTable, Design,
    name.prefix = "sample")
}
\arguments{
  \item{CountsTable}{count table}

  \item{Design}{design ( sample id / Condition / Batch )}

  \item{name.prefix}{a name for the generated plot files}
}
\value{
  Heatmap plot file adn PCA plot file in png format
}
\description{
  Plot heatmap and 2D scatter plot of principal components
}

