\name{require.auto}
\alias{require.auto}
\title{auto install of packages from Cran / Bioconductor and their dependecies (if they are not installed)
Adapted from http://sbamin.com/2012/11/05/tips-for-working-in-r-automatically-install-missing-package/}
\usage{
  require.auto(x)
}
\arguments{
  \item{x}{package to be installed}
}
\description{
  auto install of packages from Cran / Bioconductor and
  their dependecies (if they are not installed) Adapted
  from
  http://sbamin.com/2012/11/05/tips-for-working-in-r-automatically-install-missing-package/
}

